gh-2498: Upgrade store integration tests (and everything else) to use JUnit 5 (#2648)
* Fix hdfs-library to use only JUnit 5

* Fix operation to use only JUnit 5

* Fix road-traffic-generators to use only JUnit 5

* Update integration-test impl test suite tests (except loader) to use JUnit 5

* Refactor AbstractStore to use JUnit 5

* Replace ParameterizedLoaderIT - Parameters cannot be used on classes with JUnit 5

* Refactor integration test data loaders to work with JUnit 5
Nested tests which repeat the same tests with different inputs are used to replace where Parameters were used in JUnit 4. This is because Parameters are not supported at the class level in JUnit 5.

* Refactor Suite ITs to use built in JUnit 5 annotations

* Remove JUnit 4 dependencies, add JUnit 5 suite dependency
Also rename version property for JUnit to not have the version - as only 5 is used now

* Update Copyright headers

* Checkstyle fixes

* Another checkstyle fix

* Add Suites dependency to integration-test only
Adding globally caused an issue with surefire in spring-rest

* Manage dependency version centrally, introduce where required

* Copyright Update

* Replace dummy test initialisation with BeforeAllCallback Extension and parameter annotation
The dummy test initialisation is not guaranteed to run before the ITs, which causes tests to be skipped. Replaced it with an extension and improved error handling to fail tests if config is missing - instead of skipping/not running which creates a false impression of passing tests.

* Copyright Header Update

* Ensure integration tests which must be extended are not run standalone

* Copyright