Gh-2201: Add support for Accumulo 2 while retaining Accumulo 1 (#2679)
* Make Gaffer compatible with Accumulo 2 and retain support for Accumulo 1 via a utility class
Add utility class to dynamically import Accumulo classes which are incompatible between versions 1 & 2, and handle creation of certain objects accociated with these classes.
Gaffer can be built for Accumulo 1 and Hadoop 2 using a 'classic' profile which swaps dependency versions.
Other minor POM changes were required to ensure compatibility for existing libraries.

* Add and update Copyright Headers

* Remove code for handling Mock Accumulo
MiniAccumuloCluster has replaced Mock Accumulo and so this code is no longer used or required

* Use AssertJ for LegacySupport tests

* Improve modified RFileReaderIteratorTest to use root cause

* Checkstyle

* Use reflection to invoke writer/reader builder
Fixes a bytecode inconsistency that prevents portabity to both versions of Accumulo. If reflection is not used the bytecode will refer to a specific method and return type (for this builder it differs between v1/v2 of Accumulo), causing the bytecode to change when compiled against v1/v2 of Accumulo which breaks portability.

* Bump versions of Hadoop & Guava

* Add 'classic' classifiers for wars and shaded/spring-exec jars
This ensures 'classic' versions of these files do not conflict with the default versions

* Remove redundant war dependeny declarations

* Add enforcer rule to block building with JDK 11 and old Hadoop

* Update version details in Accumulo README

* Rename profile from 'classic' to 'legacy'

* Update iterators shaded JAR to work with Accumulo 2
In Accumulo 2 classes from Jackson are used, in the previous version Jackson wasn't used at all. A different version to Gaffer is used and this causes a conflict. This is fixed by relocating the class in the iterators shaded JAR. In addition, other classes have be relocated to avoid similar issues. Log4J has been removed from the shaded JAR because Accumulo already suppies this.

* Update docs on iterators JAR installation to cover Accumulo 2
Also fix example properties file - this should give the FQCN

* Replace String value with Boolean for if statements