Gh-2591: Fix String test errors on Windows (#2714)
* Remove hardcoded UNIX line returns in store tests
Replaced with platform agnostic version

* Remove redundant regex path check
This check is not required. If the path is invalid an InvalidPathException will be thrown by Java when trying to convert it. This regex duplicates that and isn't compatible with Windows paths or relative paths.

* Replace hardcoded UNIX newlines

* Add a check to convert path with backslashes to use forwardslashes
This is required because backslashes are interpreted as escaped chars.

* Restore check for path validity, but use Path object instead of regex
Also change invalid path to one which really is invalid and change variable naem to 'file' instead of 'final'.

* Update Copyright Headers

* Use String.format with '%n' platform-specific line separator
This approach has been used before in #197 to fix similar problems

* fix newline at end of file removed by mistake