Gh-2581: Replace getTraits usages with operation (#2901)
* Remove getTraits from Store

* Remove getTraits from Graph

* Remove getTraits from Map Store

* Remove getTraits from Accumulo Store

* Remove getTraits from Core Rest

* Remove getTraits from Proxy Store

* Fix to ensure user info doesn't get lost
Federated Store needs the user

* Remove getTraits from Federated Store

* Remove getTraits from Spring Rest

* Revert disabling IT for Federated Store
Fails due to QUERY_AGGREGATION Trait not being reported as a Trait of this store

* Change uses of GetTraits operation to get all supported traits
By default this operation gets the currently supported traits. It needs to get all supported instead as that is what the method used to do.

* Improve description of endpoint for Spring

* Simplify test in FederatedStoreUtilTest

* Fix for handling null User in OperationChainValidator

* Update Copyright

* Restore getTraits method for Map Store internal use
Using the static field directly would work, except this is overriden in some child classes which requires a getter to be used.

* Enable handler for GetTraits in proxy store
Previously this was unused as the handler was not added. Handler works the same as getTrait method did, removing visibility trait from returned traits. Also set other unused handler to null.

* Copyright corrections

* Fix test to include newly supported Operation in Proxy Store

* ProxyStoreBasicIT should get all supported traits and not current
This has no effect on the test currently but makes it clear what should be tested

* Revert superseded change identified in PR comment